stdlib, sim: merged LoopPointManager.py& removed setup
Merged LoopPointManager.py into LoopPoint.py and removed the setup
function.
Changed the UInt64 to Addr Param type in the SimObject file.
Now BaseLoopPoint is responsible for connecting all the LoopPoint
information with all the running cores.
