--- src/cpu/probes/LoopPointManager.py
@@ -1,49 +1,0 @@
-# Copyright (c) 2022 The Regents of the University of California
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are
-# met: redistributions of source code must retain the above copyright
-# notice, this list of conditions and the following disclaimer;
-# redistributions in binary form must reproduce the above copyright
-# notice, this list of conditions and the following disclaimer in the
-# documentation and/or other materials provided with the distribution;
-# neither the name of the copyright holders nor the names of its
-# contributors may be used to endorse or promote products derived from
-# this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-# A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-# OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-# SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-# LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-# DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-# THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-from m5.params import *
-from m5.objects import SimObject
-
-class LoopPointManager(SimObject):
-
-    type = "LoopPointManager"
-    cxx_header = "cpu/probes/looppointmanager.hh"
-    cxx_class = "gem5::LoopPointManager"
-
-    target_count = VectorParam.Int("the target PC count")
-    target_pc = VectorParam.UInt64("the target PC")
-    region_id = VectorParam.Int("the simulation region id")
-    relative_pc = VectorParam.UInt64("the relative PC of the target PC")
-    relative_count = VectorParam.Int("the relative PC count of the target PC")
-
-    def setup(self, cpulist):
-        """This function takes in a list of AbstractCore and connect a 
-        LoopPoint to each of them.
-        
-        :param cpulist : a list of AbstractCore
-        """
-        for core in cpulist:
-            core.addLoopPointProbe(self.target_pc+self.relative_pc, self)
